@@ -11,32 +11,32 @@
     "prettier": "prettier --write \"src/**/*.{ts,tsx}\""
   },
   "dependencies": {
-    "axios": "^0.26.1",
+    "axios": "^1.3.4",
     "babel": "^6.23.0",
     "eslint-config-prettier": "^8.7.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "leaflet": "^1.7.1",
+    "leaflet": "^1.9.3",
     "prettier": "^2.8.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-leaflet": "^3.2.5"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.2.1"
   },
   "devDependencies": {
     "@rollup/plugin-babel": "^6.0.3",
     "@rollup/plugin-commonjs": "^24.0.1",
     "@rollup/plugin-node-resolve": "^15.0.1",
     "@types/leaflet": "^1.9.1",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.15.0",
-    "@typescript-eslint/parser": "^5.15.0",
-    "@vitejs/plugin-react": "^1.0.7",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.54.1",
+    "@typescript-eslint/parser": "^5.54.1",
+    "@vitejs/plugin-react": "^3.1.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^8.36.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react": "^7.32.2",
     "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.5.4",
-    "vite": "^2.8.0"
+    "typescript": "^4.9.5",
+    "vite": "^4.1.4"
   }
 }
